requirements.txt
# nnU-Net framework
nnunetv2>=2.0

# Geospatial libraries
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0

# HTTP requests
requests>=2.31.0

# Deep learning (already included with nnunetv2 but explicit for reference)
# torch>=2.0.0
# torchvision>=0.15.0

# Optional utilities
tqdm>=4.66.0

